Moved files. Renamed files.
- Moved `smart_pot` into `lib/implementations`
- Renamed `EEPROM_promter` to `EEPROM_prompter`
- Moved `EEPROM_prompter` and `monitor` into `lib/utils`
