+++ lib/implementations/smart_pot.c
